requirements.txt
matplotlib==3.5.2
numpy==1.23.0
opencv_python==4.6.0.66
pandas==1.4.3
Pillow==9.5.0
pydicom==2.3.1
PyYAML==6.0
scikit_learn==1.2.2
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
